+++ src/types/PositionTypes.sol
@@ -17,13 +17,13 @@ struct BasePositionParams {
 
 /// @notice Parameters specific to full-range positions
 struct FullRangeParams {
-    uint256 tokenAmount;
-    uint256 currencyAmount;
+    uint128 tokenAmount;
+    uint128 currencyAmount;
 }
 
 /// @notice Parameters specific to one-sided positions
 struct OneSidedParams {
-    uint256 amount;
+    uint128 amount;
     bool inToken;
 }
